+++ ct_core/include/ct/core/systems/ControlledSystem-impl.h
@@ -8,13 +8,13 @@ Licensed under the BSD-2 license (see LICENSE file in main directory)
 namespace ct {
 namespace core {
 
-template <typename MANIFOLD, size_t CONTROL_DIM, bool CONT_T>
+template <typename MANIFOLD, int CONTROL_DIM, bool CONT_T>
 ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::ControlledSystem(const SYSTEM_TYPE& type)
     : System<MANIFOLD, CONT_T>(type), controller_(nullptr)
 {
 }
 
-template <typename MANIFOLD, size_t CONTROL_DIM, bool CONT_T>
+template <typename MANIFOLD, int CONTROL_DIM, bool CONT_T>
 ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::ControlledSystem(std::shared_ptr<Controller_t> controller,
     const SYSTEM_TYPE& type)
     : System<MANIFOLD, CONT_T>(type), controller_(controller)
@@ -22,7 +22,7 @@ ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::ControlledSystem(std::shared_pt
     controlAction_.setZero();
 }
 
-template <typename MANIFOLD, size_t CONTROL_DIM, bool CONT_T>
+template <typename MANIFOLD, int CONTROL_DIM, bool CONT_T>
 ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::ControlledSystem(const ControlledSystem& arg)
     : System<MANIFOLD, CONT_T>(arg), controlAction_(arg.controlAction_)
 {
@@ -30,30 +30,30 @@ ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::ControlledSystem(const Controll
         controller_ = std::shared_ptr<Controller_t>(arg.controller_->clone());
 }
 
-template <typename MANIFOLD, size_t CONTROL_DIM, bool CONT_T>
+template <typename MANIFOLD, int CONTROL_DIM, bool CONT_T>
 ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::~ControlledSystem()
 {
 }
 
-template <typename MANIFOLD, size_t CONTROL_DIM, bool CONT_T>
+template <typename MANIFOLD, int CONTROL_DIM, bool CONT_T>
 void ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::setController(const std::shared_ptr<Controller_t>& controller)
 {
     controller_ = controller;
 }
 
-template <typename MANIFOLD, size_t CONTROL_DIM, bool CONT_T>
+template <typename MANIFOLD, int CONTROL_DIM, bool CONT_T>
 void ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::getController(std::shared_ptr<Controller_t>& controller) const
 {
     controller = controller_;
 }
 
-template <typename MANIFOLD, size_t CONTROL_DIM, bool CONT_T>
+template <typename MANIFOLD, int CONTROL_DIM, bool CONT_T>
 auto ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::getController() -> std::shared_ptr<Controller_t>
 {
     return controller_;
 }
 
-template <typename MANIFOLD, size_t CONTROL_DIM, bool CONT_T>
+template <typename MANIFOLD, int CONTROL_DIM, bool CONT_T>
 void ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::computeDynamics(const MANIFOLD& m,
     const Time_t& t,
     typename MANIFOLD::Tangent& derivative)
@@ -66,7 +66,7 @@ void ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::computeDynamics(const MANI
     computeControlledDynamics(m, t, controlAction_, derivative);
 }
 
-template <typename MANIFOLD, size_t CONTROL_DIM, bool CONT_T>
+template <typename MANIFOLD, int CONTROL_DIM, bool CONT_T>
 auto ControlledSystem<MANIFOLD, CONTROL_DIM, CONT_T>::getLastControlAction() const -> control_vector_t
 {
     return controlAction_;
